pyproject.toml
[project]
name = "rosdomofon"
version = "0.1.5"
description = "ORM для работы с API РосДомофон"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.28.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.1.1",
    "kafka-python>=2.0.0"
]


[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true


[project.urls]
Homepage = "https://github.com/darkClaw921/rosdomofon"
Documentation = "https://github.com/darkClaw921/rosdomofon"
Repository = "https://github.com/darkClaw921/rosdomofon.git"
"Bug Tracker" = "https://github.com/darkClaw921/rosdomofon/issues"
Changelog = "https://github.com/darkClaw921/rosdomofon/blob/master/CHANGELOG.md"
